requirements.txt
pydantic-settings==2.9.1
pyarrow==15.0.2
pandas==1.5.3
tqdm==4.66.2
sentence-transformers==2.6.1
transformers==4.40.1
loguru==0.7.2
python-dotenv==1.0.1
fastapi==0.115.9
uvicorn==0.29.0
typing-extensions==4.9.0


langchain==0.3.0
faiss-cpu==1.7.4
langchain-community==0.3.25
langchain-huggingface==0.3.0
huggingface-hub==0.30.2


langchain-groq
slowapi==0.1.5
spacy~=3.6.0